requirements.txt
flask
python-telegram-bot
fuzzywuzzy
python-Levenshtein
gunicorn